Move CustomizeYourProductScreen to its own directory
