+++ lib/ui/screens/customize_your_product_screen/CustomizeYourProductScreen.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:sellio_mobile/core/design_system/themes/sellio_theme_provider.dart';
 import 'package:sellio_mobile/core/design_system/widgets/sellio_app_bar.dart';
-
-import '../../core/design_system/constants/assets.dart';
+import '../../../core/design_system/constants/assets.dart';
 
 class CustomizeYourProductScreen extends StatefulWidget {
   const CustomizeYourProductScreen({super.key});
@@ -12,7 +10,8 @@ class CustomizeYourProductScreen extends StatefulWidget {
       _CustomizeYourProductScreenState();
 }
 
-class _CustomizeYourProductScreenState extends State<CustomizeYourProductScreen> {
+class _CustomizeYourProductScreenState
+    extends State<CustomizeYourProductScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -21,7 +20,7 @@ class _CustomizeYourProductScreenState extends State<CustomizeYourProductScreen>
         showLeading: false,
         subtitle: "Choose a product to customize it ",
         actionIcon: Assets.orderIcon,
-      )
+      ),
     );
   }
 }
