+++ lib/ui/screens/main_screen/MainScreen.dart
@@ -4,7 +4,7 @@ import '../CartScreen.dart';
 import '../home/home_screen.dart';
 import '../AccountScreen.dart';
 import '../ThriftScreen.dart';
-import '../CustomizeYourProductScreen.dart';
+import '../customize_your_product_screen/CustomizeYourProductScreen.dart';
 
 class MainScreen extends StatefulWidget {
   final int screenIndex;
